pyproject.toml
[project]
name = "bandaid"
version = "0.1.0"
description = "A Python program that automatically detects a person's arm in a photo and digitally places a band-aid on it."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.6.1",
    "mediapipe>=0.10.21",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "pytest>=8.4.2",
]

[project.scripts]
bandaid = "bandaid.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
